pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markdown-convert"
version = "0.1.0"
description = "Convert PDF files to Markdown with OCR support"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Ezaz Ul Haq", email = "your.email@example.com"}
]
keywords = ["pdf", "markdown", "converter", "ocr", "tesseract", "pymupdf"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Utilities",
]

dependencies = [
    "pymupdf>=1.23.0",
    "pymupdf4llm>=0.0.10",
    "pytesseract>=0.3.10",
    "Pillow>=9.0.0",
    "fitz>=0.0.1.dev2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/ezazulhaq/markdown-convert"
Repository = "https://github.com/ezazulhaq/markdown-convert"
Issues = "https://github.com/ezazulhaq/markdown-convert/issues"

[project.scripts]
markdown-convert = "markdown_convert.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
markdown_convert = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py38"
select = ["E", "F", "I", "N", "W"]
ignore = []
